requirements.txt
argcomplete==1.8.2
astroid==2.2.5
beautifulsoup4==4.5.3
chardet==2.3.0
docx2txt==0.6
EbookLib==0.15
isort==4.3.13
lazy-object-proxy==1.3.1
lxml==4.3.2
mccabe==0.6.1
nltk==3.4
pdf2image==1.4.2
Pillow==5.4.1
pylint==2.3.1
PyPDF2==1.26.0
pytesseract==0.2.7
python-pptx==0.6.5
singledispatch==3.4.0.3
six==1.10.0
SpeechRecognition==3.6.3
typed-ast==1.3.1
Wand==0.5.1
wrapt==1.11.1
xlrd==1.0.0
XlsxWriter==1.1.5
